// nms/app/theme/default.d.ts
/**
 * Copyright 2022 The Magma Authors.
 *
 * This source code is licensed under the BSD-style license found in the
 * LICENSE file in the root directory of this source tree.
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export const colors: {
  primary: {
    white: string;
    selago: string;
    concrete: string;
    mercury: string;
    nobel: string;
    gullGray: string;
    comet: string;
    brightGray: string;
    mirage: string;
  };
  secondary: {
    malibu: string;
    dodgerBlue: string;
    mariner: string;
  };
  button: {
    lightOutline: string;
    fill: string;
  };
  state: {
    positive: string;
    positiveAlt: string;
    error: string;
    errorAlt: string;
    errorFill: string;
    warning: string;
    warningAlt: string;
    warningFill: string;
  };
  alerts: {
    severe: string;
    major: string;
    minor: string;
    other: string;
  };
  data: {
    coral: string;
    flamePea: string;
    portage: string;
    studio: string;
  };
  code: {
    crusta: string;
    pelorous: string;
    electricViolet: string;
    orchid: string;
    chelseaCucumber: string;
    candlelight: string;
    mischka: string;
  };
};

export const shadows: {
  DP1: string;
  DP2: string;
  DP3: string;
  DP4: string;
  DP5: string;
};

export const typography: {
  h1: any;
  h2: any;
  h3: any;
  h4: any;
  h5: any;
  h6: any;
  subtitle1: any;
  subtitle2: any;
  subtitle3: any;
  body1: any;
  body2: any;
  body3: any;
  code: any;
  button: any;
  caption: any;
  overline: any;
};

export default any;